Node.js: Getting Started by Samer Buna (Pluralsight)
Link: https://app.pluralsight.com/library/courses/nodejs-getting-started/table-of-contents
Web and Mobile applications depend on back-end services to host their various resources and do most of the heavy lifting. Creating these back-end services is often very difficult, but the Node.js runtime makes it easier and more fun. In this course, Node.js: Getting Started, you'll learn the basic skills you need to create scalable back-end services with Node.js. First, you'll learn about the tools that come built-in with Node as well as some of the tools that are offered by the community. Next, you'll learn about package management, module dependencies, event-driven concurrency strategies, and how to write code to interface with the operating systems on servers. Finally, you'll discover the modern JavaScript features that you can use with Node.js today. When you're finished with this course, you'll be comfortable creating and executing code for Node.